[🔨 Refactor] NoteView Navigation 경로 삭제
